refactor: move task dtos to task folder
