+++ src/application/usecases/tasks/create-task.ts
@@ -1,6 +1,6 @@
 import { Task } from "../../../domain/entities/task"; 
 import { TaskRepository } from "../../../domain/repositories/task.repository"; 
-import { CreateTaskDto } from "../../dtos/create-task.dto"; 
+import { CreateTaskDto } from "../../dtos/task/create-task.dto"; 
 
 export class CreateTask {
   constructor(private readonly taskRepository: TaskRepository) {}
